# backend/handlers/auth.go
package handlers

import (
	"context"
	"encoding/json"
	"net/http"
	"os"
	"time"

	"handbook/config"
	"handbook/models"

	"github.com/golang-jwt/jwt/v5"
	"go.mongodb.org/mongo-driver/bson"
	"go.mongodb.org/mongo-driver/bson/primitive"
	"golang.org/x/crypto/bcrypt"
)

// Register godoc
// @Summary Register a new user
// @Description create a new user account
// @Tags auth
// @Accept  json
// @Produce  json
// @Param user body models.User true "User registration info"
// @Success 201 {object} map[string]string
// @Router /register [post]
func Register(w http.ResponseWriter, r *http.Request) {
	var input models.User
	if err := json.NewDecoder(r.Body).Decode(&input); err != nil {
		http.Error(w, "Invalid input", http.StatusBadRequest)
		return
	}

	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
	defer cancel()

	var existing models.User
	err := config.DB.Collection("users").FindOne(ctx, bson.M{"email": input.Email}).Decode(&existing)
	if err == nil {
		http.Error(w, "Email already exists", http.StatusBadRequest)
		return
	}

	hash, _ := bcrypt.GenerateFromPassword([]byte(input.Password), bcrypt.DefaultCost)
	input.Password = string(hash)
	input.Role = "user"

	_, err = config.DB.Collection("users").InsertOne(ctx, input)
	if err != nil {
		http.Error(w, "Error creating user", http.StatusInternalServerError)
		return
	}

	w.WriteHeader(http.StatusCreated)
	json.NewEncoder(w).Encode(map[string]string{"message": "User created"})
}

// Login godoc
// @Summary Login user
// @Description authenticate user and return JWT
// @Tags auth
// @Accept  json
// @Produce  json
// @Param credentials body models.User true "User credentials"
// @Success 200 {object} map[string]string
// @Router /login [post]
func Login(w http.ResponseWriter, r *http.Request) {
	var input models.User
	if err := json.NewDecoder(r.Body).Decode(&input); err != nil {
		http.Error(w, "Invalid input", http.StatusBadRequest)
		return
	}

	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
	defer cancel()

	var user models.User
	err := config.DB.Collection("users").FindOne(ctx, bson.M{"email": input.Email}).Decode(&user)
	if err != nil {
		http.Error(w, "Invalid credentials", http.StatusUnauthorized)
		return
	}

	if err := bcrypt.CompareHashAndPassword([]byte(user.Password), []byte(input.Password)); err != nil {
		http.Error(w, "Invalid credentials", http.StatusUnauthorized)
		return
	}

	token := jwt.NewWithClaims(jwt.SigningMethodHS256, jwt.MapClaims{
		"id":  user.ID.Hex(),
		"exp": time.Now().Add(24 * time.Hour).Unix(),
	})

	tokenString, _ := token.SignedString([]byte(os.Getenv("JWT_SECRET")))

	json.NewEncoder(w).Encode(map[string]string{"token": tokenString})
}

func AuthMiddleware(next http.HandlerFunc) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		authHeader := r.Header.Get("Authorization")
		if authHeader == "" {
			http.Error(w, "Missing token", http.StatusUnauthorized)
			return
		}

		tokenString := authHeader
		if len(authHeader) > 7 && authHeader[:7] == "Bearer " {
			tokenString = authHeader[7:]
		}

		token, err := jwt.Parse(tokenString, func(token *jwt.Token) (interface{}, error) {
			return []byte(os.Getenv("JWT_SECRET")), nil
		})

		if err != nil || !token.Valid {
			http.Error(w, "Invalid token", http.StatusUnauthorized)
			return
		}

		claims, _ := token.Claims.(jwt.MapClaims)
		userID := claims["id"].(string)

		ctx := context.WithValue(r.Context(), "userID", userID)
		next.ServeHTTP(w, r.WithContext(ctx))
	}
}

// Profile godoc
// @Summary Get user profile
// @Description get current user details, their courses and enrolled courses
// @Tags auth
// @Accept  json
// @Produce  json
// @Security Bearer
// @Success 200 {object} map[string]interface{}
// @Router /profile [get]
func Profile(w http.ResponseWriter, r *http.Request) {
	userID := r.Context().Value("userID").(string)
	objID, _ := primitive.ObjectIDFromHex(userID)

	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
	defer cancel()

	var user models.User
	err := config.DB.Collection("users").FindOne(ctx, bson.M{"_id": objID}).Decode(&user)
	if err != nil {
		http.Error(w, "User not found", http.StatusNotFound)
		return
	}

	user.Password = ""

	// Fetch user's courses
	cursor, err := config.DB.Collection("courses").Find(ctx, bson.M{"author_id": objID})
	var courses []models.Course
	if err == nil {
		cursor.All(ctx, &courses)
	}
	if courses == nil {
		courses = []models.Course{}
	}

	// Fetch enrolled courses
	progCursor, err := config.DB.Collection("progress").Find(ctx, bson.M{"user_id": objID})
	var progressList []models.Progress
	if err == nil {
		progCursor.All(ctx, &progressList)
	}

	var enrolledCourses []map[string]interface{}
	for _, p := range progressList {
		var c models.Course
		err := config.DB.Collection("courses").FindOne(ctx, bson.M{"_id": p.CourseID}).Decode(&c)
		if err == nil {
			enrolledCourses = append(enrolledCourses, map[string]interface{}{
				"course":   c,
				"progress": p,
			})
		}
	}
	if enrolledCourses == nil {
		enrolledCourses = []map[string]interface{}{}
	}

	response := map[string]interface{}{
		"user":             user,
		"courses":          courses,
		"enrolled_courses": enrolledCourses,
	}

	json.NewEncoder(w).Encode(response)
}